Install ibus component declaration file globally

+++ launcher.sh
@@ -3,11 +3,17 @@
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 cd $DIR
 
-# Update the installation
-#
-# ATTENTION: This command will delete all your local changes.
-git reset --hard HEAD
-git pull
+if [ $1 == "--ibus" ]
+then
+	# Update the installation
+	#
+	# ATTENTION: This command will delete all your local changes.
+	git reset --hard HEAD
+	git pull
 
-python3 ibus_engine/main.py
+	python3 ibus_engine/main.py --ibus
+else
+	python3 ibus_engine/main.py
+fi
+
